web: split SidebarPinButton into a standalone component
