+++ web/src/SidebarResources.test.tsx
@@ -7,7 +7,8 @@ import { LocalStorageContextProvider, makeKey } from "./LocalStorage"
 import PathBuilder from "./PathBuilder"
 import SidebarItem from "./SidebarItem"
 import { SidebarItemBox } from "./SidebarItemView"
-import { SidebarPinButton, SidebarPinContextProvider } from "./SidebarPin"
+import { SidebarPinContextProvider } from "./SidebarPin"
+import SidebarPinButton from "./SidebarPinButton"
 import SidebarResources, { SidebarListSection } from "./SidebarResources"
 import { twoResourceView } from "./testdata"
 import { ResourceView } from "./types"
